pyproject.toml
[project]
name = "prompt-generator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.0",
    "python-dotenv>=1.0.1",
]
